refactor(core): move IM bot providers to services
