+++ src/crates/services/services-integrations/src/remote_connect/bot/mod.rs
@@ -4,19 +4,23 @@
 //! concrete session/runtime dependencies are reduced to service ports.
 
 mod command;
+pub mod feishu;
 mod locale;
 mod menu;
 mod state;
+pub mod telegram;
 
 use serde::{Deserialize, Serialize};
 
 pub use command::{parse_command, BotCommand};
+pub use feishu::{FeishuBotApi, FeishuConfig};
 pub use locale::{fmt_count, strings_for, BotLanguage, BotStrings};
 pub use menu::{MenuItem, MenuItemStyle, MenuView};
 pub use state::{
     BotAction, BotActionStyle, BotChatState, BotDisplayMode, BotInteractionHandler,
     BotInteractiveRequest, BotMessageSender, BotQuestion, BotQuestionOption, PendingAction,
 };
+pub use telegram::{TelegramBotApi, TelegramConfig};
 
 /// Configuration for a bot-based connection.
 #[derive(Debug, Clone, Serialize, Deserialize)]
